Rename default.lua to config.lua
