+++ lua/smoothcursor/callbacks/default.lua
@@ -1,16 +1,12 @@
-local buffer = require('smoothcursor.callbacks').buffer
-local is_enabled = require('smoothcursor.callbacks').is_enabled
-local sc_timer = require('smoothcursor.callbacks').sc_timer
-local replace_signs = require('smoothcursor.callbacks').replace_signs
-local place_sign = require('smoothcursor.callbacks').place_sign
-local unplace_signs = require('smoothcursor.callbacks').unplace_signs
-local fancy_head_exists = require('smoothcursor.callbacks').fancy_head_exists
-local config = require('smoothcursor.default')
+local callback = require('smoothcursor.callbacks')
+local buffer = callback.buffer
+
+local config = require('smoothcursor.config').config
 local debug_callback = require('smoothcursor.debug').debug_callback
 
 -- Default corsor callback. buffer["prev"] is always integer
 local function sc_default()
-  if not is_enabled() then
+  if not callback.is_enabled() then
     return
   end
   local cursor_now = vim.fn.getcurpos(vim.fn.win_getid())[2]
@@ -21,9 +17,9 @@ local function sc_default()
   buffer['diff'] = math.min(buffer['diff'], vim.fn.winheight(0) * 2)
   buffer['w0'] = vim.fn.line('w0')
   buffer['w$'] = vim.fn.line('w$')
-  if math.abs(buffer['diff']) > config.default_args.threshold then
+  if math.abs(buffer['diff']) > config.threshold then
     local counter = 1
-    sc_timer:post(function()
+    callback.sc_timer:post(function()
       cursor_now = vim.fn.getcurpos(vim.fn.win_getid())[2]
       if buffer['prev'] == nil then
         buffer['prev'] = cursor_now
@@ -36,31 +32,31 @@ local function sc_default()
       buffer['diff'] = buffer['prev'] - cursor_now
       buffer['prev'] = buffer['prev']
         - (
-          (buffer['diff'] > 0) and math.ceil(buffer['diff'] / 100 * config.default_args.speed)
-          or math.floor(buffer['diff'] / 100 * config.default_args.speed)
+          (buffer['diff'] > 0) and math.ceil(buffer['diff'] / 100 * config.speed)
+          or math.floor(buffer['diff'] / 100 * config.speed)
         )
       buffer:push_front(buffer['prev'])
       -- Replace Signs
-      replace_signs()
+      callback.replace_signs()
       counter = counter + 1
       debug_callback(buffer, { 'Jump: True' })
       -- Timer management
       if
-        counter > (config.default_args.timeout / config.default_args.intervals)
+        counter > (config.timeout / config.intervals)
         or (buffer['diff'] == 0 and buffer:is_stay_still())
       then
-        if not fancy_head_exists() then
-          unplace_signs()
+        if not callback.fancy_head_exists() then
+          callback.unplace_signs()
         end
-        sc_timer:abort()
+        callback.sc_timer:abort()
       end
     end)
   else
     buffer['prev'] = cursor_now
     buffer:all(cursor_now)
-    unplace_signs()
-    if fancy_head_exists() then
-      place_sign(buffer['prev'], 'smoothcursor')
+    callback.unplace_signs()
+    if callback.fancy_head_exists() then
+      callback.place_sign(buffer['prev'], 'smoothcursor')
     end
     debug_callback(buffer, { 'Jump: False' })
   end
