+++ lua/smoothcursor/utils.lua
@@ -1,7 +1,7 @@
 local sc = {}
 local smoothcursor_started = false
 local callback = require('smoothcursor.callbacks')
-local configs = require('smoothcursor.default').default_args
+local configs = require('smoothcursor.config').config
 local init = require('smoothcursor.init')
 local buffer_leaved = false
